refactor: replace instrument bar with transport-only bar
The fullscreen profile's metric tile strip duplicated every value the
draggable readout card already floats over the chart, and on a phone it
wrapped into four rows that ate a third of the screen (#811).

Removing the tiles takes the customize sheet, the tile preference
helpers and ReadoutTile with them. What remains is a thin container
around ProfileTransportControls, so ProfileTransportBar no longer needs
Riverpod, nor the analysis and tankPressures arguments that existed only
to populate tiles.
