--- lib/features/dive_log/presentation/widgets/readout_tile.dart
@@ -1,54 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// A single instrument readout: small uppercase label over a bold value.
-///
-/// Renders an em dash when [value] is null so the instrument bar keeps a
-/// stable layout through data gaps during playback.
-class ReadoutTile extends StatelessWidget {
-  final String label;
-  final String? value;
-  final Color? valueColor;
-
-  const ReadoutTile({
-    super.key,
-    required this.label,
-    required this.value,
-    this.valueColor,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final textTheme = Theme.of(context).textTheme;
-
-    return Container(
-      constraints: const BoxConstraints(minWidth: 72),
-      padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 6),
-      decoration: BoxDecoration(
-        color: colorScheme.surfaceContainerHighest,
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            label.toUpperCase(),
-            style: textTheme.labelSmall?.copyWith(
-              color: colorScheme.onSurfaceVariant,
-              letterSpacing: 0.8,
-            ),
-          ),
-          Text(
-            value ?? '—',
-            style: textTheme.titleMedium?.copyWith(
-              fontWeight: FontWeight.w700,
-              color: valueColor ?? colorScheme.onSurface,
-              fontFeatures: [const FontFeature.tabularFigures()],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
